增加历史订单详情界面，历史订单本地数据库缓存，设置界面

+++ driver_pro/app/src/main/java/com/guokrspace/dududriver/net/SocketClient.java
@@ -411,7 +411,7 @@ public class SocketClient {
         return ret;
     }
 
-    public int getHistoryOrders (String type, int num, int orderId, ResponseHandler handler) {
+    public int getHistoryOrders (String type, int num, Long orderId, ResponseHandler handler) {
         int ret = -1;
         JSONObject params = new JSONObject();
         try {
